requirements.txt
######## requirements.txt #######
pynetworktables
PySimpleGUI
py2exe
#
